third_party/blink/web_tests/platform/generic/external/wpt/html/canvas/element/text-styles/2d.text.draw.space.collapse.start-expected.txt

This is a testharness.js-based test.
FAIL Space characters at the start of a line are collapsed (per CSS) assert_approx_equals: Red channel of the pixel at (25, 25) expected 0 +/- 2 but got 255
Harness: the test ran to completion.
